Add the VS 2008 project file for MagoRemote.

+++ branches/mago64/DebugEngine/MagoRemote/App.cpp
@@ -20,8 +20,8 @@ enum
     ConnectTimeoutErrorCode = -1,
 };
 
-UINT32 gMainTid;
-UINT32 gStartedExclusiveSession;
+UINT32  gMainTid;
+LONG    gStartedExclusiveSession;
 
 
 namespace Mago
